+++ src/include/cpu/amd/microcode.h
@@ -2,6 +2,6 @@
 #define CPU_AMD_MICROCODE_H
 
 void update_microcode(u32 cpu_deviceid);
-void amd_update_microcode_from_cbfs(u32 equivalent_processor_rev_id);
+void amd_update_microcode_from_cbfs(u16 equivalent_processor_rev_id);
 
 #endif /* CPU_AMD_MICROCODE_H */
